requirements.txt
Pyrebase4==4.6.0
requests==2.24.0
cvzone==1.5.6
opencv-python-headless==4.7.0.72
adafruit-io==2.7.0
face-recognition==1.3.0
websockets==11.0.3
requests-toolbelt==0.10.1
mediapipe==0.9.1.0